Adopt Mathieu's xor trick, extending to 64 bits as well as 32.

+++ CodeSamples/defer/rcu_nest32.h
@@ -24,7 +24,7 @@
 DEFINE_SPINLOCK(rcu_gp_lock);
 #define RCU_GP_CTR_BOTTOM_BIT 0x80000000
 #define RCU_GP_CTR_NEST_MASK (RCU_GP_CTR_BOTTOM_BIT - 1)
-int rcu_gp_ctr = 0;	/* increment by RCU_GP_CTR_BOTTOM_BIT each gp. */
+long rcu_gp_ctr = 0;	/* increment by RCU_GP_CTR_BOTTOM_BIT each gp. */
 DEFINE_PER_THREAD(long, rcu_reader_gp);
 
 static inline int rcu_old_gp_ongoing(int t)
